CoatesvilleHUB
Chester County is presented with a timely opportunity to take a bold step into the future of  business and education technology. The digital developmental plan originated from a Coatesville Community Educational Foundation report that outlines a successful future for Coatesville and the region based on a community-based internet network. The purpose of the proposed Keystone Connect Network of Coatesville is to create a regional broadband network whose purpose is to increase educational opportunities and generate business growth.